created enemy spawn with enemy animations

+++ Assets/Scripts/Enemy/EnemyConfig.cs
@@ -7,11 +7,11 @@ namespace Enemy
     public class EnemyConfig
     {
         [SerializeField] private Enemy _prefab;
-        [SerializeField, Range(1, 10)] private float _health;
-        [SerializeField, Range(1, 10)] private float _speed;
+        [SerializeField, Range(1, 10)] private int _health;
+        [SerializeField, Range(1, 10)] private int _speed;
 
         public Enemy Prefab => _prefab;
-        public float Health => _health;
-        public float Speed => _speed;
+        public int Health => _health;
+        public int Speed => _speed;
     }
 }
